ERC1155 standard

+++ contracts/week4/Topic1/DepositFunds.sol
@@ -2,14 +2,14 @@
 pragma solidity ^0.8.7;
 
 contract DepositFunds {
-    mapping(address => uint) public balances;
+    mapping(address => uint256) public balances;
 
     function deposit() public payable {
         balances[msg.sender] += msg.value;
     }
 
     function withdraw() public {
-        uint bal = balances[msg.sender];
+        uint256 bal = balances[msg.sender];
         require(bal > 0);
 
         (bool sent, ) = msg.sender.call{value: bal}("");
@@ -19,7 +19,7 @@ contract DepositFunds {
     }
 
     //   function to check the balance of this contract
-    function getBalance() public view returns (uint) {
+    function getBalance() public view returns (uint256) {
         return address(this).balance;
     }
 }
